Tests for Delete Post Endpoint
NOT done. Making Pull Request for help purposes lol. I got stuck on the not post owner. I get user already exists as an error or a too many attempts from this IP wait for a while error. Not sure if that's because I need to make a change to the register.