Delete Post Tests / Fix
Changes - create a post and check for no session token because before when creating post I got session token error. Try to delete post. Respond with 403. It actually worked, though I had to get ai help for most of this. Tried to just move it in front of the is valid test but got other errors due to  order of middleware i'm assuming.